requirement.txt
Django >=5.0,<6.0
djangorestframework>=3.14
requests
django-cors-headers
pyyaml